fem: fix a signedness comparison in the vector assembly cell loop
`x_dofmap.extent(1)` is an unsigned index type, so comparing it against
an `std::int32_t` loop counter trips -Wsign-compare. The C++ test build
does not instantiate `assemble_cells` with a dynamic geometry-dofmap
extent and so did not catch it; the Python bindings do, and failed to
build.

+++ cpp/dolfinx/fem/assemble_vector_impl.h
@@ -129,7 +129,7 @@ void assemble_cells(
     // with a run-time length that lowers to a memmove call, per vertex.
     const std::int32_t* xdofs
         = x_dofmap_ptr + static_cast<std::ptrdiff_t>(c) * ndofs_x;
-    for (std::int32_t i = 0; i < ndofs_x; ++i)
+    for (std::size_t i = 0; i < ndofs_x; ++i)
     {
       const U* src
           = x_ptr + static_cast<std::ptrdiff_t>(xdofs[i]) * x.extent(1);
